Initial set up
This app will allow users to search products and find where it is sold based on a location at a venu. Users will be able to make an account and save their favorite drinks/ food so that they can skip the process of redoing a search for common items and be able to see if it will be availible during said events.