Models added as well as .gitignore
The database flow has been defined per customer specs. There are models for menus and favorites to create a relation between stands and items as well as users to items. The latter will be used to track how many users favorite an item to display a rank next to it.